requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
google-genai==1.30.0
pdfplumber==0.11.4
python-dotenv==1.0.1
pydantic==2.9.2